feat[data-loader]: Make staging workspace non-lazy
