+++ src/Docker/Runner/DataLoader/BaseDataLoaderFactory.php
@@ -5,11 +5,9 @@ declare(strict_types=1);
 namespace Keboola\DockerBundle\Docker\Runner\DataLoader;
 
 use Keboola\StagingProvider\Staging\File\LocalStaging;
+use Keboola\StagingProvider\Staging\StagingClass;
 use Keboola\StagingProvider\Staging\StagingProvider;
 use Keboola\StagingProvider\Staging\StagingType;
-use Keboola\StagingProvider\Staging\Workspace\LazyWorkspaceStaging;
-use Keboola\StagingProvider\Staging\Workspace\WorkspaceStagingInterface;
-use Keboola\StagingProvider\Workspace\ProviderConfig\ExistingWorkspaceConfig;
 use Keboola\StagingProvider\Workspace\WorkspaceProvider;
 use LogicException;
 use Psr\Log\LoggerInterface;
@@ -27,27 +25,22 @@ abstract class BaseDataLoaderFactory
         StagingType $stagingType,
         ?string $stagingWorkspaceId,
     ): StagingProvider {
-        if ($stagingWorkspaceId === null) {
-            $stagingWorkspace = null;
-        } else {
-            $stagingWorkspace = new LazyWorkspaceStaging(
-                $this->workspaceProvider,
-                new ExistingWorkspaceConfig($stagingWorkspaceId),
-            );
-        }
-
-        $stagingProvider = new StagingProvider(
-            $stagingType,
-            $stagingWorkspace,
-            new LocalStaging($this->dataDirectory),
-        );
-
-        if ($stagingProvider->getTableDataStaging() instanceof WorkspaceStagingInterface xor
-            $stagingWorkspace !== null
+        if ($stagingType->getStagingClass() === StagingClass::Workspace xor
+            $stagingWorkspaceId !== null
         ) {
             throw new LogicException('Staging workspace ID must be configured for component with workspace staging.');
         }
 
-        return $stagingProvider;
+        if ($stagingWorkspaceId === null) {
+            $stagingWorkspace = null;
+        } else {
+            $stagingWorkspace = $this->workspaceProvider->getExistingWorkspace($stagingWorkspaceId);
+        }
+
+        return new StagingProvider(
+            $stagingType,
+            $stagingWorkspace,
+            new LocalStaging($this->dataDirectory),
+        );
     }
 }
